==> R Scripts/0. Patient Characteristic Table.Rmd <==
---
title: "0. Patient Characteristic Table"
author: "Avril Metcalfe-Roach"
date: "`r Sys.Date()`"
output: html_document
---

```{r}
library(phyloseq)
library(tidyverse)

# Load metagenomic data
ps = readRDS("../Reference Files/taxonomy_phyloseq.rds") %>% 
  subset_taxa(Species=='s__Blautia_obeum') %>% psmelt() %>% 
  select(-OTU,-Abundance,-c(Kingdom:Species))

# Other variables included in analysis
crp = read.csv('../Reference Files/serum_crp.csv') # C reactive protein
# Metabolites that are bacterially associated and correlate with disease
met = read.csv('../Reference Files/normalized_proteolytic_metabolites.csv') 
risk = read.csv('../Reference Files/Misc PD Risk.csv') # Other clinical variables of interest in PD

# Recode variables
df = ps %>% left_join(crp) %>% left_join(met) %>% left_join(risk) %>% 
  select(-sample_Sample,-redcap_event_name,-entacapone,-contains('nona'),
         -depth,-Log2CRP,-logcoffeeperday) %>% 
  rename(`Disease Duration` = Disease.duration, `Levodopa Equivalent Dose` = levo.eq.dose,
         `Laxative Use` = laxatives, `Bristol Stool Rating` = bristol, 
         `History of Smoking` = smokeprev_or_before_pd,`Coffee (Cups/Day)` = coffee_per_day,
         `Beck Depression Inventory (/63)` = bdi_total, `Fatigue Severity Scale (/7)` = fss_total,
         `MDS-UPDRS 1` = mds1.total, `MDS-UPDRS 2` = mds2.total, `MDS-UPDRS 3` = mds3.total, 
         `MDS-UPDRS 4` = mds4.total, `Total MDS-UPDRS` = mds.total,
         `C reactive protein (mg/L)` = CRP_mgdl, `p-Cresol` = pcresol, Phenylacetylglutamine = phenylacetylglutamine)
```

```{r}
# Categorical variables
df.cat = df %>% select(Sample,Status,Sex,`Laxative Use`,`History of Smoking`)
cat = tibble()
chi = tibble()
# Save N, perform chi square analyses
for(i in names(df.cat)[3:ncol(df.cat)]){
  # i = 'Sex'
  t = table(Status = df.cat$Status, test = df.cat[[i]]) 
  c = chisq.test(t) %>% broom::tidy()
  cat = rbind(cat,t %>% broom::tidy() %>% mutate(Var=i))
  chi = rbind(chi,c %>% mutate(Var=i))
}
chi$Var[chi$Var=='History of Smoking'] = 'History of Smoking (n, % Yes)'
chi$Var[chi$Var=='Laxative Use'] = 'Laxative Use (n, % Yes)'
chi$Var[chi$Var=='Sex'] = 'Sex (n, % Male)'

cat2 = cat %>% 
  mutate(test = ifelse(test=='0','No',ifelse(test=='1','Yes',test))) %>% 
  arrange(test) %>% 
  group_by(Status,Var) %>% 
  mutate(pct = round(100*n[2L]/(n[1L]+n[2L]),1),
         total = n[1L]+n[2L]) %>%
  filter(test == test[2L]) %>% ungroup %>%
  mutate(Var = paste(Var,' (n, % ',test,')',sep='')) %>% 
  select(-test) %>% 
  mutate(npct = paste0(n,' (',round(pct,1),'%)')) %>% 
  select(-n,-pct) %>% rename(n=total)
  
# Numerical variables
df.num = df %>% select(-all_of(names(df.cat)[3:ncol(df.cat)])) %>% 
  pivot_longer(cols = -c(Sample,Status), names_to = 'Var', values_to = 'value') %>% 
  filter(!is.na(value)) %>% 
  group_by(Status,Var) %>% add_count() %>% ungroup %>% 
  group_by(Status,Var,n) %>% 
  summarize(meansd = paste0(round(mean(value),1), ' (',round(sd(value),1),')'))
# Wilcoxon tests
wil = df %>% select(-all_of(names(df.cat)[3:ncol(df.cat)])) %>% 
  select(-contains('UPDRS'),-all_of(c('Levodopa Equivalent Dose','Disease Duration'))) %>% 
  pivot_longer(cols = -c(Sample,Status), names_to = 'Var', values_to = 'value') %>% 
  filter(!is.na(value)) %>%
  group_by(Var) %>%
  group_modify(~wilcox.test(value~Status,data=.) %>% broom::tidy())

# Combined
lvls = tibble(Var = c('Sex (n, % Male)','Age',
         'History of Smoking (n, % Yes)',"Coffee (Cups/Day)",
         'Laxative Use (n, % Yes)','Bristol Stool Rating',
         "C reactive protein (mg/L)",
         'p-Cresol','Phenylacetylglutamine',
         "Beck Depression Inventory (/63)","Fatigue Severity Scale (/7)",
         'Disease Duration','Levodopa Equivalent Dose',
         "MDS-UPDRS 1","MDS-UPDRS 2","MDS-UPDRS 3","MDS-UPDRS 4","Total MDS-UPDRS"),
         Group = c(rep('Demographics',4),rep('GI',2),rep('Metabolites',3),rep('Mood',2),
                   rep('PD Data',7)))
df.all = cat2 %>% full_join(df.num) %>% 
  mutate(Var = factor(Var, levels = lvls$Var)) %>% arrange(Var) %>% 
  left_join(lvls) %>% select(Group,everything()) %>% 
  mutate(Statistic = ifelse(is.na(npct),meansd,npct)) %>% 
  select(-npct,-meansd)
df.all$Statistic[str_detect(df.all$Var,'UPDRS') & df.all$Status=='Ctrl'] = NA
df.all$n[str_detect(df.all$Var,'UPDRS') & df.all$Status=='Ctrl'] = NA
df.all = df.all %>% 
  pivot_wider(names_from = Status, values_from = c(n,Statistic)) %>% 
  select(Group,Var,contains('PD'),everything())

  # Add statistics and edit names
df.all2 = df.all %>% left_join(rbind(chi %>% select(Var,p.value),wil %>% select(Var,p.value)))
df.all2 = df.all2 %>% rename(Variable = Var, `n (PD)` = n_PD, `n (Ctrl)` = n_Ctrl, `Statistic (PD)` = Statistic_PD, `Statistic (Ctrl)` = Statistic_Ctrl, `P value` = p.value)
```

```{r}
writexl::write_xlsx(df.all2, '../Results/Table S1 - Patient Characteristic Table.xlsx')
```
