==> R Scripts/2. Differentially Abundant Taxa - Fig. 2a.Rmd <==
---
title: "2. Differentially Abundant Taxa - Fig. 2a"
author: "Avril Metcalfe-Roach"
date: "`r Sys.Date()`"
output: html_document
---

```{r}
library(phyloseq)
library(ggpubr)
library(data.table)
library(tidyverse)

# This contains all the taxonomic and sample data necessary for cross-sectional analyses.
ps = readRDS("../Reference Files/taxonomy_phyloseq.rds")
```

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

```{r Plot MetaCyc Pathways}
# Sig Species
df1 = readxl::read_xlsx('../Results/Supplementary Data/Differential Abundance/Taxonomy_Univar_Stats.xlsx',sheet = 'Species') %>% 
  dplyr::select(Species,DA, association_dir, q_val) %>% rename(Direction= association_dir) %>% 
  pivot_wider(names_from = DA, values_from = c(Direction,q_val)) %>% 
  mutate(`# Tools` = ((`q_val_ANCOM-BC`<0.05) + (q_val_ALDEx2<0.05) + (q_val_Maaslin2<0.05))) %>% 
  mutate(`Sig (Univar)` = `# Tools` >1) %>% 
  filter(`Sig (Univar)`==T)

df2 = readxl::read_xlsx('../Results/Supplementary Data/Differential Abundance/Taxonomy_Multivar_Stats.xlsx',sheet = 'Species') %>% 
  filter(Variable == 'StatusPD') %>% 
  dplyr::select(Species,DA, association_dir, q_val) %>% rename(Direction= association_dir) %>% 
  pivot_wider(names_from = DA, values_from = c(Direction,q_val)) %>% 
  mutate(`Sig (Model)` = ((`q_val_ANCOM-BC`<0.05) + (q_val_ALDEx2<0.05) + (q_val_Maaslin2<0.05) >1))

df = df1 %>% left_join(df2 %>% dplyr::select(Species, `Sig (Model)`))

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Calculate log 2 fc
lfc = ps %>% microbiome::transform('compositional') %>% psmelt() %>% 
  filter(Species %in% df$Species) %>%
  # Add alpha value depending on whether it's univariately or multivariately sig
  mutate(alpha = Species %in% df$Species[df$`Sig (Model)`==T]) %>% 
  group_by(Status,Species,alpha) %>% 
  summarize(Abundance = mean(Abundance),.groups='keep') %>% ungroup %>% 
  pivot_wider(names_from = Status, values_from = Abundance) %>% 
  mutate(lfc = log2(PD/Ctrl)) %>% arrange(lfc)
lfc$Species = factor(lfc$Species, levels = lfc$Species)

# Plot 1: LFC
p1 = lfc %>% mutate(COL = ifelse(lfc<0,'#01BEC3','#FC746E')) %>% 
  ggplot(aes(lfc,Species,fill = COL, alpha = alpha)) +
  geom_col(aes(fill = COL)) +
  theme_minimal(base_size = 16) +
  scale_alpha_discrete(range = c(0.35,1)) +
  ylab(NULL) + xlab('Log2 FC (PD/Ctrl)') +
  theme(legend.position = 'none')
p1

# Plot 2: CLR-transformed abundances
p2 = ps %>% microbiome::transform('clr') %>% psmelt() %>% 
  filter(Species %in% df$Species) %>%
  # Add alpha value depending on whether it's univariately or multivariately sig
  mutate(alpha = Species %in% df$Species[df$`Sig (Model)`==T]) %>% 
  mutate(Species = factor(Species, levels = lfc$Species)) %>% 
  ggplot(aes(Abundance,Species,fill = Status, alpha = alpha)) +
  geom_boxplot() +
  scale_alpha_discrete(range = c(0.35,1)) +
  theme_minimal(base_size = 16) +
  theme(axis.text.y = element_blank()) +
  theme(legend.position = 'none') +
  xlab('CLR') + ylab(NULL)

plot.out = ggarrange(plotlist = list(p1+theme(legend.position='none'),p2),
                     common.legend = F, ncol=2, widths = c(3,1.7))
ggsave('../Results/2. Differential Abundance & Enrichments/Fig. 2a - Differentially Abundant Species.png', height = 5.5, width =10)
```
