==> R Scripts/4a. Key Species Per EC.Rmd <==
---
title: "4a. Key Species Per EC"
author: "Avril Metcalfe-Roach"
date: "`r Sys.Date()`"
output: html_document
---

```{r}
library(phyloseq)
library(data.table)
library(ggpubr)
library(tidyverse)
library(gt)

# Phyloseq object of MetaCyc pathways
ps = readRDS('../Reference Files/functional_phyloseq_ec_stratified.rds')
```

```{r Select pathways of interest}
# Load significant pwys
sig_uni = readxl::read_xlsx('../Results/Supplementary Data/Differential Abundance/EC_Stats.xlsx',sheet='univar') %>% 
  filter(q_val<0.05) %>% group_by(Species) %>% add_count() %>% filter(n>1) %>% 
  select(Species,association_dir) %>% unique
sig_multi = readxl::read_xlsx('../Results/Supplementary Data/Differential Abundance/EC_Stats.xlsx',sheet='multivar') %>%
  filter(Variable=='StatusPD', Species %in% sig_uni$Species) %>% 
  group_by(DA) %>% mutate(q_val = p.adjust(p_val,method='BH')) %>% ungroup %>% 
  filter(q_val<0.05) %>% group_by(Species) %>% add_count() %>% filter(n>1) %>% 
  select(Species,association_dir) %>% unique

# Identify OTUs that are >10% prevalent
p = apply(ps@otu_table@.Data, 1, function(x) (sum(x != 0, na.rm = TRUE)/length(x[!is.na(x)]))>0.1)
p2 = rownames(ps@tax_table@.Data)[p]

# Normalized data for lfc calculations & stats analysis
df2 = ps %>% microbiome::transform('clr')
df2 = prune_taxa(p2,df2) %>% # prevalence
  subset_taxa(feature %in% sig_multi$Species) %>% psmelt()
df3 = ps %>% microbiome::transform('compositional') 
df3 = prune_taxa(p2,df3) %>% # prevalence
  subset_taxa(feature %in% sig_multi$Species) %>% psmelt()
```

# Run Analysis

```{r Run and format}
# Univariate correlations
uni = df2 %>% group_by(feature,taxon) %>% 
  group_modify(~summary(glm(Abundance~Status, data = .))$coefficients %>% 
                 as.matrix() %>% as.data.frame() %>% rownames_to_column('var')) %>% 
  ungroup() %>% rename(pval = `Pr(>|t|)`) %>% 
  filter(!str_detect(var,'Intercept')) %>% 
  group_by(var) %>% 
  mutate(qval = p.adjust(pval,method = 'fdr'), .after = pval) %>% 
  ungroup() %>% 
  mutate(Type = 'Univariate') %>% 
  # Add association direction
  left_join(sig_uni %>% rename(feature=Species)) %>% 
  mutate(same_direction = (Estimate<0) == (association_dir=='Negative'))

# Multivariate correlations
mult = df2 %>% group_by(feature,taxon) %>% 
  group_modify(~summary(glm(Abundance~Status+Sex+laxatives+depth, data = .))$coefficients %>% 
                 as.matrix() %>% as.data.frame() %>% rownames_to_column('var')) %>% 
  ungroup() %>% rename(pval = `Pr(>|t|)`) %>% 
  filter(!str_detect(var,'Intercept')) %>% 
  group_by(var) %>%
  mutate(qval = p.adjust(pval,method = 'fdr'), .after = pval) %>% 
  ungroup() %>% 
  mutate(Type = 'Multivariate') %>% 
  # Add association direction
  left_join(sig_uni %>% rename(feature=Species)) %>% 
  mutate(same_direction = (Estimate<0) == (association_dir=='Negative'))

# Filter to include only q<0.05, and only disease status-related rows for multivar
unisig = uni %>% filter(qval<0.05, same_direction==T)
multisig = mult %>% filter(qval<0.05, same_direction==T, var=='StatusPD')

# Filter to only include disease status hits that are significant in both uni and multivar results
sig_filt = multisig %>% filter(paste0(feature,taxon) %in% paste0(unisig$feature,unisig$taxon)) %>% 
  arrange(pval) %>% arrange(qval)
```
```{r}
# Hierarchy data - extract only EC terms that are present in dataset
# EC and EC2 are the same, except that EC2 doesn't have -RXN suffixes
brite = read.csv('../Reference Files/EC_key.csv')
# Remove any instances where there are multiple ECs listed for a single entry. Since only the first 3 numbers matter, just use the first version of every instance as the first 3 are always the same.
# Update dataset names
brite = brite %>% mutate(ec = sapply(.$ec, function(x) str_split(x,' //')[[1]][1]))
# Load EC key
ec_key = read.csv('../Reference Files/EC_key_classes.csv') %>% 
  mutate(type = paste(Class,Subclass,Subsubclass,sep = '.'))

# Add multivar stats for all pathways that are univariately significant
stats = rbind(unisig, 
              mult %>% filter(var=='StatusPD',
                              paste0(feature,taxon) %in% paste0(unisig$feature,unisig$taxon))) %>% 
  rename(Pathway = feature,Species=taxon,Variable = var) %>% 
  pivot_wider(values_from = c(Estimate:qval), names_from = Type) %>% 
  mutate(Sig = (qval_Univariate<0.05 & qval_Multivariate < 0.05))


stats.edit = stats %>% 
  mutate(Pathway = sapply(.$Pathway,function(x) x %>% str_split(':') %>% 
                            .[[1]] %>% .[1] %>% str_remove('-RXN'))) %>% 
  mutate(Pathway = sapply(.$Pathway, function(x){
    if(x %in% brite$EC2){
      return(brite$ec[which(x==brite$EC2)])
    } else {return(x)}
    })) %>% 
  mutate(type = sapply(.$Pathway,function(x){
    s = str_split(x,'[.]') %>% .[[1]]
    if(length(s)>3){
      return(s[-length(s)] %>% paste(collapse = '.'))
    } else{
      return(x)
    } 
      })) %>% 
  mutate(type=as.character(type)) %>% 
  left_join(ec_key) %>% 
  select(Pathway,contains('name'),Species,everything()) %>% 
  select(-c(type:Subsubclass)) %>% 
  rename(EC = Pathway,Class=class.name, Subclass = subclass.name, Subsubclass = subsubclass.name)
  

# More formatting
gt_tbl = stats.edit %>% select(EC, everything()) %>%
  mutate(Species = str_remove(Species,'.*[.]') %>% str_remove_all('s__') %>% 
           str_replace_all('_',' ') %>% str_replace('uncl','Uncl')) %>% 
  mutate_if(is.numeric,signif,3) %>% 
  mutate_at(vars(contains('pval'),contains('qval')),scales::scientific,3) %>% 
  arrange(qval_Univariate) %>% arrange(-Sig) %>% arrange(Species) %>% arrange(EC)
# Even more formatting
gt2 = gt_tbl %>% filter(is.na(Species)) %>%  # empty tibble, just colnames
  rbind(matrix(data=c('Reduced in PD',rep('',ncol(gt_tbl)-1)),nrow=1, dimnames=list(NULL,names(gt_tbl)))) %>% 
  rbind(gt_tbl %>% filter(association_dir=='Negative')) %>%
  rbind(matrix(data=c('Elevated in PD',rep('',ncol(gt_tbl)-1)),nrow=1, dimnames=list(NULL,names(gt_tbl)))) %>% 
  rbind(gt_tbl %>% filter(association_dir=='Positive')) %>% 
  select(-association_dir,-Variable) %>% 
  unique
```


```{r Save raw statistics}
writexl::write_xlsx(list(Univariate = uni, Multivariate = mult, Sig = gt2),
                    '../Results/Supplementary Data/Differential Abundance/Stratified EC.xlsx')
```
