==> R Scripts/7b. Build Models from Random Forest Analysis.Rmd <==
---
title: "7b. Build Models from Random Forest Analysis"
author: "Avril Metcalfe-Roach"
date: "`r Sys.Date()`"
output: html_document
---

```{r}
library(phyloseq)
library(ggpubr)
library(data.table)
library(tidyverse)
library(readxl)
library(scales)
```

# Importance Values - Status

```{r Formatting function and dataframes}
format_features = function(v){
  temp = sapply(v,function(x){
    if(str_detect(x,'s__')){
      a = x %>% str_remove('s__') %>% str_split('_') %>% .[[1]]
      a[1] = paste(str_sub(a[1],end=1),'.')
      return(paste(a[1],a[2]))
    } else {
      return(str_wrap(x,20))
    }
  })
  return(temp)
}
fix = tibble(Feature = c('Sex','smokeprev_or_before_pd','bristol','logcoffeeperday'),
             fix = c('Sex (Male)', "Ever Smoked",'Bristol Rating','Coffee/Day'))
meta_names = tibble(Feature = c("PWY-7383: anaerobic energy metabolism (invertebrates, cytosol)",
                                "PWY-6471: peptidoglycan biosynthesis IV (Enterococcus faecium)",
                                "PWY66-409: superpathway of purine nucleotide salvage",
                                "PWY-5177: glutaryl-CoA degradation",     
                                "PWY-7242: D-fructuronate degradation",     
                                "GLUCUROCAT-PWY: superpathway of &beta;-D-glucuronide and D-glucuronate degradation",
                                "GALACT-GLUCUROCAT-PWY: superpathway of hexuronide and hexuronate degradation"),
                    new = c("PWY-7383:\nanaerobic metabolism",
                                "PWY-6471:\npeptidoglycan biosynth.",
                                "PWY66-409:\npurine salvage",
                                "PWY-5177:\nglutaryl-CoA degrad.",     
                                "PWY-7242:\nfructuronate degrad.",     
                                "GLUCUROCAT-PWY:\nglucuronide/ate degrad.",
                                "GALACT-GLUCUROCAT-PWY:\nhexuronide/ate degrad."))
key = read.csv('../Reference Files/EC_key.csv') %>% rename(Feature=EC) %>% select(-EC2) %>% 
  mutate(ec = sapply(ec, function(x) x %>% str_split(' // ') %>% .[[1]] %>% .[1]))
```

```{r Formatted Importance Values}
# Calculate Directions for each variable (PD vs Ctrl)
d1 = readxl::read_xlsx('../Results/Supplementary Data/Differential Abundance/Taxonomy_Multivar_Stats.xlsx',sheet = 'Species') %>%
  filter(Variable == 'StatusPD') %>% 
  select(Species,association_dir) %>% unique()
d2 = readxl::read_xlsx('../Results/Supplementary Data/Differential Abundance/MetaCyc_Stats.xlsx',sheet='multivar') %>% 
  filter(Variable=='StatusPD') %>% 
  select(Species,association_dir) %>% unique()
d3 = readxl::read_xlsx('../Results/Supplementary Data/Differential Abundance/KO_Stats.xlsx',sheet='multivar') %>% 
  filter(Variable=='StatusPD') %>% 
  select(Species,association_dir) %>% unique()
d4 = readxl::read_xlsx('../Results/Supplementary Data/Differential Abundance/COG Enzymes_Stats.xlsx',sheet='multivar') %>% 
  filter(Variable=='StatusPD') %>% 
  select(Species,association_dir) %>% unique()
d5 = readxl::read_xlsx('../Results/Supplementary Data/Differential Abundance/EC_Stats.xlsx',sheet='multivar') %>% 
  filter(Variable=='StatusPD') %>% 
  select(Species,association_dir) %>% unique()

ts7 = read.csv('../Reference Files/bacterial_metabolites.csv')
df = read.csv('../Reference Files/FOR RF - Status - Metabolites - Annotated Only.csv', check.names = F) %>% 
  mutate(Status = factor(Status),Sex = as.numeric(Sex=='Male')) %>% 
  pivot_longer(cols=any_of(ts7$Metabolite.Identity),names_to = 'met', values_to = 'value') %>% 
  filter(!is.na(value)) %>% 
  group_by(met) %>% 
  group_modify(~summary(glm(Status~value,family='binomial',data=.))$coefficients %>% as.matrix %>% 
                 as.data.frame %>% rownames_to_column('Variable')) %>% 
  filter(Variable!='(Intercept)') %>% 
  mutate(association_dir = ifelse(Estimate==0,'Neutral',ifelse(Estimate<0,"Ctrl",'PD'))) %>% 
  select(met,association_dir) %>% rename(Species=met)
cov = c('bristol','Sex','logcoffeeperday','smokeprev_or_before_pd')
dfcov = read.csv('../Reference Files/FOR RF - Status - Taxonomy.csv', check.names = F) %>% 
  mutate(Status = factor(Status),Sex = as.numeric(Sex=='Male'),
         smokeprev_or_before_pd = as.numeric(smokeprev_or_before_pd=='Yes')) %>% 
  pivot_longer(cols=all_of(cov),names_to = 'met', values_to = 'value') %>% 
  filter(!is.na(value)) %>% 
  group_by(met) %>% 
  group_modify(~summary(glm(Status~value,family='binomial',data=.))$coefficients %>% as.matrix %>% 
                 as.data.frame %>% rownames_to_column('Variable')) %>% 
  filter(Variable!='(Intercept)') %>% 
  mutate(association_dir = ifelse(Estimate==0,'Neutral',ifelse(Estimate<0,"Ctrl",'PD'))) %>% 
  select(met,association_dir) %>% rename(Species=met)

# All associations with Status
dir = rbind(d1,d2,d3,d4,d5,df,dfcov)

# Load Importance Values
imp = read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Status - Covars.xlsx') %>% mutate(Data='Covar') %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Status - Taxonomy.xlsx', 
                      sheet='Taxonomy') %>% mutate(Data='Taxonomy')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Status - Taxonomy.xlsx', 
                      sheet='Taxonomy & Covar') %>% mutate(Data='Taxonomy & Covar')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Status - MetaCyc.xlsx', 
                      sheet='MetaCyc') %>% mutate(Data='MetaCyc')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Status - MetaCyc.xlsx', 
                      sheet='MetaCyc & Covar') %>% mutate(Data='MetaCyc & Covar')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Status - KO.xlsx', 
                      sheet='KO') %>% mutate(Data='KO')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Status - KO.xlsx', 
                      sheet='KO & Covar') %>% mutate(Data='KO & Covar')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Status - COG.xlsx', 
                      sheet='COG') %>% mutate(Data='COG')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Status - COG.xlsx', 
                      sheet='COG & Covar') %>% mutate(Data='COG & Covar')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Status - EC.xlsx', 
                      sheet='EC') %>% mutate(Data='EC')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Status - EC.xlsx', 
                      sheet='EC & Covar') %>% mutate(Data='EC & Covar')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Status - Metabolites Annot&Covar.xlsx', 
                      sheet='Metabolites') %>% mutate(Data='Metabolites')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Status - Metabolites Annot&Covar.xlsx', 
                      sheet='Metabolites & Covar') %>% mutate(Data='Metabolites & Covar'))

# Add direction info and format the Feature names for plotting
imp2 = imp %>% left_join(dir %>% rename(Feature = Species)) %>% 
  left_join(fix) %>% mutate(Feature = ifelse(!is.na(fix),fix,Feature)) %>% 
  left_join(meta_names) %>% mutate(Feature = ifelse(!is.na(new),new,Feature)) %>% 
  mutate(Feature = format_features(.$Feature)) 
imp2$Feature[str_detect(imp2$Data,'EC|COG|KO')] = sapply(imp2$Feature[str_detect(imp2$Data,'EC|COG|KO')],
                                                            function(x) x %>% str_split(':') %>% .[[1]] %>% .[1])
imp2 = imp2 %>% mutate(Feature = ifelse(str_detect(Data,'EC|COG|KO'), str_remove_all(Feature,' |\n'),Feature))
imp2 = imp2 %>% 
  left_join(key) %>% 
  mutate(Feature = ifelse(!is.na(ec),ec,Feature)) %>% 
  select(-ec,-new,-fix) %>% 
  rename(`PD vs Ctrl` = association_dir) %>% 
  mutate(`PD vs Ctrl` = ifelse(`PD vs Ctrl` == 'Positive','PD',
                               ifelse(`PD vs Ctrl` == 'Negative','Ctrl',`PD vs Ctrl`)))

# If it's control associated, change the Importance value to a negative
imp3 = imp2 %>% mutate(Importance = ifelse(`PD vs Ctrl`=='Ctrl',Importance*-1,Importance)) %>% 
  arrange(Importance) %>% 
  mutate(Data = factor(Data, levels = c('Covar',sort(unique(.$Data))[sort(unique(.$Data))!='Covar']))) %>% 
  arrange(Data) %>% 
  select(-`...1`)

# For reference below:
imp4 = imp2 %>% rename(fixed = Feature) %>% left_join(imp)
```

```{r Save Importance Table}
L = list(Taxonomy = imp3 %>% filter(str_detect(Data,'Taxonomy') | Data=='Covar'),
         MetaCyc = imp3 %>% filter(str_detect(Data,'MetaCyc') | Data=='Covar'),
         KO = imp3 %>% filter(str_detect(Data,'KO') | Data=='Covar'),
         COG = imp3 %>% filter(str_detect(Data,'COG') | Data=='Covar'),
         EC = imp3 %>% filter(str_detect(Data,'EC') | Data=='Covar'),
         Metabolites = imp3 %>% filter(str_detect(Data,'Metabolites') | Data=='Covar'))

writexl::write_xlsx(L,'../Results/Supplementary Data/Random Forest/00. RF Importance Values - Status - With Directions.xlsx')
```

```{r Plot Importance Values}
hex <- scales::hue_pal()(2) %>% rev()

# Taxonomy
p1 = imp3 %>% filter(Data=='Covar') %>% mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`PD vs Ctrl`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=18) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab('Importance') +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
p2 = imp3 %>% filter(Data=='Taxonomy') %>% mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`PD vs Ctrl`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=18) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab(NULL) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
p3 = imp3 %>% filter(Data=='Taxonomy & Covar') %>% mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`PD vs Ctrl`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=18) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab(NULL) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggarrange(p1,p2,p3, ncol=3, widths = c(4,6,8))

ggsave('../Results/5. Random Forest/Importance Scores - Status - Taxonomy.jpeg',height=3.5, width=11)

# MetaCyc
p4 = imp3 %>% filter(Data=='MetaCyc') %>% 
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`PD vs Ctrl`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab(NULL) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
p5 = imp3 %>% filter(Data=='MetaCyc & Covar') %>%
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`PD vs Ctrl`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab(NULL) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggarrange(p4,p5, ncol=2, widths = c(6,8))

ggsave('../Results/5. Random Forest/Importance Scores - Status - MetaCyc.jpeg',height=4, width=11)

# KO
p4 = imp3 %>% filter(Data=='KO') %>% 
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`PD vs Ctrl`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab('Importance') +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
p5 = imp3 %>% filter(Data=='KO & Covar') %>%
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`PD vs Ctrl`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab('Importance') +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggarrange(p4,p5, nrow=2)

ggsave('../Results/Supplementary Figures/Importance Scores - Status - KO.jpeg',height=6, width=11)

# EC
p4 = imp3 %>% filter(Data=='EC') %>%
  .[-which(duplicated(.$Feature)),] %>% 
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`PD vs Ctrl`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab('Importance') +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
p5 = imp3 %>% filter(Data=='EC & Covar') %>%
  .[-which(duplicated(.$Feature)),] %>% 
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`PD vs Ctrl`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab('Importance') +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggarrange(p4,p5, nrow=2)

ggsave('../Results/Supplementary Figures/Importance Scores - Status - EC.jpeg',height=6, width=11)

# COG
p4 = imp3 %>% filter(Data=='COG') %>% 
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`PD vs Ctrl`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab('Importance') +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
p5 = imp3 %>% filter(Data=='COG & Covar') %>%
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`PD vs Ctrl`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab('Importance') +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggarrange(p4,p5, nrow=2)

ggsave('../Results/Supplementary Figures/Importance Scores - Status - COG.jpeg',height=6, width=11)

# Metabolites
p4 = imp3 %>% filter(Data=='Metabolites') %>% 
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`PD vs Ctrl`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab('Importance') +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
p5 = imp3 %>% filter(Data=='Metabolites & Covar') %>%
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`PD vs Ctrl`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab('Importance') +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggarrange(p4,p5, nrow=2)

ggsave('../Results/Supplementary Figures/Importance Scores - Status - Metabolites.jpeg',height=8, width=11)
```

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# Importance Values - Progression

```{r Formatting function and dataframes}
format_features = function(v){
  temp = sapply(v,function(x){
    if(str_detect(x,'s__')){
      a = x %>% str_remove('s__') %>% str_split('_') %>% .[[1]]
      a[1] = paste(str_sub(a[1],end=1),'.')
      return(paste(a[1],a[2]))
    } else {
      return(str_wrap(x,20))
    }
  })
  return(temp)
}
fix = tibble(Feature = c('Sex','smokeprev_or_before_pd','bristol','logcoffeeperday'),
             fix = c('Sex (Male)', "Ever Smoked",'Bristol Rating','Coffee/Day'))
meta_names = tibble(Feature = c("PWY-7383: anaerobic energy metabolism (invertebrates, cytosol)",
                                "PWY-6471: peptidoglycan biosynthesis IV (Enterococcus faecium)",
                                "PWY66-409: superpathway of purine nucleotide salvage",
                                "PWY-5177: glutaryl-CoA degradation",     
                                "PWY-7242: D-fructuronate degradation",     
                                "GLUCUROCAT-PWY: superpathway of &beta;-D-glucuronide and D-glucuronate degradation",
                                "GALACT-GLUCUROCAT-PWY: superpathway of hexuronide and hexuronate degradation"),
                    new = c("PWY-7383:\nanaerobic metabolism",
                                "PWY-6471:\npeptidoglycan biosynth.",
                                "PWY66-409:\npurine salvage",
                                "PWY-5177:\nglutaryl-CoA degrad.",     
                                "PWY-7242:\nfructuronate degrad.",     
                                "GLUCUROCAT-PWY:\nglucuronide/ate degrad.",
                                "GALACT-GLUCUROCAT-PWY:\nhexuronide/ate degrad."))
key = read.csv('../Reference Files/EC_key.csv') %>% rename(Feature=EC) %>% select(-EC2) %>% 
  mutate(ec = sapply(ec, function(x) x %>% str_split(' // ') %>% .[[1]] %>% .[1]))
```

```{r Formatted Importance Values}
# Calculate Directions for each variable (PD vs Ctrl)

d1 = readxl::read_xlsx('../Results/Supplementary Data/Progression/Species vs Disease Progression Stats.xlsx',sheet = 'Multivar') %>%
  filter(term=='months_last_visit:Abundance',Variable=='mds4.total') %>% 
  select(Species,estimate) %>% unique() %>% 
  mutate(association_dir = ifelse(estimate<0,"Slow","Fast")) %>% select(-estimate)
d2 = readxl::read_xlsx('../Results/Supplementary Data/Progression/MetaCyc vs Disease Progression Stats.xlsx',sheet = 'Multivar') %>%
  filter(term=='months_last_visit:Abundance',Variable=='mds4.total') %>% 
  select(Species,estimate) %>% unique() %>% 
  mutate(association_dir = ifelse(estimate<0,"Slow","Fast")) %>% select(-estimate)
d3 = readxl::read_xlsx('../Results/Supplementary Data/Progression/KO vs Disease Progression Stats.xlsx',sheet = 'Multivar') %>%
  filter(term=='months_last_visit:Abundance',Variable=='mds4.total') %>% 
  select(Species,estimate) %>% unique() %>% 
  mutate(association_dir = ifelse(estimate<0,"Slow","Fast")) %>% select(-estimate)
d4 = readxl::read_xlsx('../Results/Supplementary Data/Progression/COG vs Disease Progression Stats.xlsx',sheet = 'Multivar') %>%
  filter(term=='months_last_visit:Abundance',Variable=='mds4.total') %>% 
  select(Species,estimate) %>% unique() %>% 
  mutate(association_dir = ifelse(estimate<0,"Slow","Fast")) %>% select(-estimate)
d5 = readxl::read_xlsx('../Results/Supplementary Data/Progression/EC vs Disease Progression Stats.xlsx',sheet = 'Multivar') %>%
  filter(term=='months_last_visit:Abundance',Variable=='mds4.total') %>% 
  select(Species,estimate) %>% unique() %>% 
  mutate(association_dir = ifelse(estimate<0,"Slow","Fast")) %>% select(-estimate)

ts7 = read.csv('../Reference Files/bacterial_metabolites.csv')
df = read.csv('../Reference Files/FOR RF - Progression - Metabolites - Annotated Only.csv', check.names = F) %>% 
  mutate(Progression = factor(Progression, levels=c('Slow','Med','Fast')),Sex = as.numeric(Sex=='Male')) %>% 
  pivot_longer(cols=any_of(ts7$Metabolite.Identity),names_to = 'met', values_to = 'value') %>% 
  filter(!is.na(value),!is.na(Estimate)) %>% 
  group_by(met) %>% 
  group_modify(~summary(glm(Estimate~value,data=.))$coefficients %>% as.data.frame %>% rownames_to_column('var')) %>% 
  filter(var=='value') %>% 
  select(met,Estimate) %>% rename(Species=met) %>% 
  mutate(association_dir = ifelse(Estimate<0,"Slow",'Fast')) %>% 
  select(-Estimate)
cov = c('bristol','Sex','logcoffeeperday')
dfcov = read.csv('../Reference Files/FOR RF - Progression - Taxonomy.csv', check.names = F) %>% 
  mutate(Sex = as.numeric(Sex=='Male')) %>% 
  select(Sample,Estimate,all_of(cov)) %>% 
  pivot_longer(cols=all_of(cov),names_to = 'met', values_to = 'value') %>% 
  filter(!is.na(value),!is.na(Estimate)) %>% 
  group_by(met) %>% 
  group_modify(~summary(glm(Estimate~value,data=.))$coefficients %>% as.data.frame %>% rownames_to_column('var')) %>% 
  filter(var=='value') %>% 
  select(met,Estimate) %>% rename(Species=met) %>% 
  mutate(association_dir = ifelse(Estimate<0,"Slow",'Fast')) %>% 
  select(-Estimate)

# All associations with Status
dir = rbind(d1,d2,d3,d4,d5,df,dfcov)

# Load Importance Values
imp = read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Progression - Covars.xlsx') %>% mutate(Data='Covar') %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Progression - Taxonomy.xlsx', 
                      sheet='Taxonomy') %>% mutate(Data='Taxonomy')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Progression - Taxonomy.xlsx', 
                      sheet='Taxonomy & Covar') %>% mutate(Data='Taxonomy & Covar')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Progression - MetaCyc.xlsx', 
                      sheet='MetaCyc') %>% mutate(Data='MetaCyc')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Progression - MetaCyc.xlsx', 
                      sheet='MetaCyc & Covar') %>% mutate(Data='MetaCyc & Covar')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Progression - KO.xlsx', 
                      sheet='KO') %>% mutate(Data='KO')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Progression - KO.xlsx', 
                      sheet='KO & Covar') %>% mutate(Data='KO & Covar')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Progression - COG.xlsx', 
                      sheet='COG') %>% mutate(Data='COG')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Progression - COG.xlsx', 
                      sheet='COG & Covar') %>% mutate(Data='COG & Covar')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Progression - EC.xlsx', 
                      sheet='EC') %>% mutate(Data='EC')) %>% 
  full_join(read_xlsx('../Results/Supplementary Data/Random Forest/RF Importance Values - Progression - EC.xlsx', 
                      sheet='EC & Covar') %>% mutate(Data='EC & Covar'))

# Add direction info and format the Feature names for plotting
imp2 = imp %>% left_join(dir %>% rename(Feature=Species)) %>% 
  left_join(fix) %>% mutate(Feature = ifelse(!is.na(fix),fix,Feature)) %>% 
  left_join(meta_names) %>% mutate(Feature = ifelse(!is.na(new),new,Feature)) %>% 
  mutate(Feature = format_features(.$Feature)) 
imp2$Feature[str_detect(imp2$Data,'EC|COG|KO')] = sapply(imp2$Feature[str_detect(imp2$Data,'EC|COG|KO')],
                                                            function(x) x %>% str_split(':') %>% .[[1]] %>% .[1])
imp2 = imp2 %>% mutate(Feature = ifelse(str_detect(Data,'EC|COG|KO'), str_remove_all(Feature,' |\n'),Feature))
imp2 = imp2 %>% mutate(Feature = ifelse(str_detect(Feature,'ARGININE-DEIMINASE'),'ARGININE-DEIMINASE-RXN',Feature)) %>% 
  left_join(key) %>% 
  mutate(Feature = ifelse(!is.na(ec),ec,Feature)) %>% 
  select(-ec,-new,-fix) %>% 
  rename(`Slow vs Fast` = association_dir) %>% 
  mutate(`Slow vs Fast` = ifelse(`Slow vs Fast` == 'Positive','Fast',
                               ifelse(`Slow vs Fast` == 'Negative','Slow',`Slow vs Fast`)))

# If it's control associated, change the Importance value to a negative
imp3 = imp2 %>% mutate(Importance = ifelse(`Slow vs Fast`=='Slow',Importance*-1,Importance)) %>% 
  arrange(Importance) %>% 
  mutate(Data = factor(Data, levels = c('Covar',sort(unique(.$Data))[sort(unique(.$Data))!='Covar']))) %>% 
  arrange(Data) %>% 
  select(-`...1`)

# For reference below:
imp4 = imp2 %>% rename(fixed = Feature) %>% left_join(imp)
```

```{r Save Importance Table}
L = list(Taxonomy = imp3 %>% filter(str_detect(Data,'Taxonomy') | Data=='Covar'),
         MetaCyc = imp3 %>% filter(str_detect(Data,'MetaCyc') | Data=='Covar'),
         KO = imp3 %>% filter(str_detect(Data,'KO') | Data=='Covar'),
         COG = imp3 %>% filter(str_detect(Data,'COG') | Data=='Covar'),
         EC = imp3 %>% filter(str_detect(Data,'EC') | Data=='Covar'),
         Metabolites = imp3 %>% filter(str_detect(Data,'Metabolites') | Data=='Covar'))

writexl::write_xlsx(L,'../Results/Supplementary Data/Random Forest/00. RF Importance Values - Progression - With Directions.xlsx')
```

```{r Plot Importance Values}
hex <- scales::hue_pal()(2) %>% rev()

# Taxonomy
p1 = imp3 %>% filter(Data=='Covar') %>% mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`Slow vs Fast`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=18) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab('Importance') +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
p2 = imp3 %>% filter(Data=='Taxonomy') %>% mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`Slow vs Fast`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=18) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab(NULL) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
p3 = imp3 %>% filter(Data=='Taxonomy & Covar') %>% mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`Slow vs Fast`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=18) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab(NULL) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggarrange(p1,p2,p3, ncol=3, widths = c(4,6,8))

ggsave('../Results/5. Random Forest/Importance Scores - Progression - Taxonomy.jpeg',height=3.5, width=11)

# MetaCyc
p4 = imp3 %>% filter(Data=='MetaCyc') %>% 
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`Slow vs Fast`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab(NULL) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
p5 = imp3 %>% filter(Data=='MetaCyc & Covar') %>%
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`Slow vs Fast`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab(NULL) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggarrange(p4,p5, ncol=2, widths = c(6,8))

ggsave('../Results/5. Random Forest/Importance Scores - Progression - MetaCyc.jpeg',height=4, width=11)

# KO
p4 = imp3 %>% filter(Data=='KO') %>% 
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`Slow vs Fast`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab('Importance') +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
p5 = imp3 %>% filter(Data=='KO & Covar') %>%
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`Slow vs Fast`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab(NULL) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggarrange(p4,p5, nrow=2)

ggsave('../Results/Supplementary Figures/Importance Scores - Progression - KO.jpeg',height=6, width=11)

# EC
p4 = imp3 %>% filter(Data=='EC') %>%
  .[-which(duplicated(.$Feature)),] %>% 
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`Slow vs Fast`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab('Importance') +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
p5 = imp3 %>% filter(Data=='EC & Covar') %>%
  .[-which(duplicated(.$Feature)),] %>% 
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`Slow vs Fast`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab(NULL) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggarrange(p4,p5, nrow=2)

ggsave('../Results/Supplementary Figures/Importance Scores - Progression - EC.jpeg',height=6, width=11)

# COG
p4 = imp3 %>% filter(Data=='COG') %>% 
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`Slow vs Fast`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab('Importance') +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
p5 = imp3 %>% filter(Data=='COG & Covar') %>%
  mutate(Feature=factor(Feature,levels=.$Feature)) %>% 
    ggplot(aes(y = Importance, x = Feature, fill=`Slow vs Fast`)) + geom_col() +
    scale_fill_manual(values = hex) +
    theme_classic(base_size=16) +
    theme(legend.position = 'none')  +  
  scale_y_continuous(expand = expansion(mult = c(0,0.15))) +xlab(NULL) + ylab(NULL) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggarrange(p4,p5, nrow=2)

ggsave('../Results/Supplementary Figures/Importance Scores - Progression - COG.jpeg',height=6, width=11)
```
